Move model classes that are not strictly v2 API to generic package
The ModelName and Metadata classes are not strictly v2. Modelname
is also used in v1 and Metadata with its attributes described by
Dependency, NamedTypeItem, and Type are expected to be used in
the v3 protocol as well. This commits moves them from their current
location in the com.puppetlabs.geppetto.forge.v2.model package to
the com.puppetlabs.geppetto.forge.model package.