Adopt PuppetDB client changes for version v3 and allow non-SSL connect
The PuppetDB java client version 1.5 adopts the new v3 client API. It
is not backward compatible since some attribute names have changed.
This commit adds the changes needed to be able to use the new client.

The commit also removes the requirement to use SSL connections when
connecting to PuppetDB.